INSERT write customer letter salutation
Only on INSERT, if a salutation in TO Online can be found based on IN_SALUTATION, and if this salutation has a letter salutation prefix, concatenate this prefix with the surname of the customer, and write the letter salutation for the new customer. For example, if the new customer is sent with an IN_LANG = DE, the IN_SALUTATION = Herr, and the letter salutation prefix is found to be "Sehr geehrter Herr", then the new customer will be created with a letter salutation "Sehr geehrter Herr Bargholz", if the IN_LASTNAME = Bargholz.